+++ changelog.md
@@ -19,6 +19,8 @@ This is a breaking version: configs and data packs have got a huge update
   * Removed hoes cooldown
 * Stack Sizes
   * Removed stackable soups. Just use item definitions
+* Health Regen
+  * Passive regen has been moved here from No Hunger
 * Changed, moved and renamed many config options
   * Client
     * World border has its own feature now